requirements.txt
tqdm==4.66.5
onnx==1.17.0
onnx2torch==1.5.15
numpy==1.23.4
matplotlib==3.8.4
matplotlib-inline==0.1.7
mediapipe==0.10.15
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.11.4
chumpy==0.70
pillow==9.4.0
open3d==0.17.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
imageio==2.35.1
kornia==0.7.3
yacs==0.1.8
plyfile==1.1
roma==1.5.0
face-alignment==1.4.1
filterpy==1.4.5
ipython==8.27.0
jupyterlab
dearpygui==2.0.0
git+https://github.com/facebookresearch/pytorch3d.git@V0.7.8